Experimentally replace xUnit with MS Test
xUnit doesn't seem to support UWP any more in practice, so this change tries moving everything over to MSTest instead, because it continues to work on VS 2022 with UWP code.

Although this is a fairly mechanical translation (we're not using anything in xUnit's API that doesn't have direct equivalents in other test frameworks), this has uncovered one issue. Unlike xUnit, MSTest doesn't supply a SynchronizationContext, which causes several of the tests to execute difference code paths. Some of them fail as a result.

It's unclear whether this has uncovered latent bugs in Rx, or whether the absence of a synchronization context effectively makes these tests invalid. For now, since this is just an experiment, I've disabled the offending tests.