Move UWP code out of System.Reactive.
UWP is a source of many problems in the tooling, and the single biggest reason for including a UWP-specific version in the main `System.Reactive` package was that UWP used not to have full support for threads, necessitating a special UWP-flavoured ThreadPoolScheduler.

But as of Windows 10.0.16299, UWP provided full .NET Standard 2.0 support, so there's no longer any strong reason to have a UWP-specific build. (Annoyingly, some UWP-specific surface area sneaked in, which this resolves by moving that into a new WindowsThreadPoolScheduler in the System.Reactive.WindowsRuntime project. This is a breaking change for UWP clients relying on how ThreadPoolScheduler looked in v5.0, but it only requires a reference to this other package, and to change to using WindowsThreadPoolScheduler.)

All UWP-specific surface area is now in a separate library. Most of it has simply moved back where it used to be - type converters for most of the types were in place, so they've just come back home.

This commit a few project changes that are not fundamental to the goal, but which seemed to be necessary to work around problems arising from limited support for UWP in the tooling. For example, I ended up having to remove the use of MSBuild.Sdk.Extras from some projects.